chore: share integration capture harness for cli tests
